Kmer cluster set change
Kmer cluster set change

Instead of unifying the kmer sets into a new merged set after a hit is found, we instead designate the secondary set as a child of the primary.

Subsequent hits match against the parent or one its child sets.

The problem the old method is that the unified sets could get so large that it would pull matches out of thin air due to the sheer number of kmers it had available to play mix and match with.

Update align.py