Pool change
Uses a single pool of workers instead of creating a new pool for each set of arguments. This reduces the overhead of creating and tearing down the pool.

In this modified code, we create the pool outside the loop over input directories, and then reuse the same pool for each set of arguments. This can help reduce the overhead of creating and tearing down the pool for each set of arguments. Note that we still need to use a separate call to pool.starmap() for each set of arguments.